# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerRequest": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "question_id": {
            "title": "Question Id",
            "type": "string"
          },
          "scenario_patch": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Patch"
          }
        },
        "required": [
          "question_id",
          "answer"
        ],
        "title": "AnswerRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartAuditRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AuditMode"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "scenario_revision": {
            "minimum": 1.0,
            "title": "Scenario Revision",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "scenario_revision",
          "mode"
        ],
        "title": "StartAuditRequest",
        "type": "object"
      },
...
      "SuccessEnvelope_AuditResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditResponse"
          },
          "error": {
            "title": "Error",
            "type": "null"
          }
        },
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[AuditResponse]",
        "type": "object"
      },
...
      "SuccessEnvelope_list_AuditEvent__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Data",
            "type": "array"
          },
          "error": {
            "title": "Error",
            "type": "null"
          }
        },
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[list[AuditEvent]]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PolygraphML API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audits": {
      "post": {
        "operationId": "start_audit_api_v1_audits_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Audit"
      }
    },
    "/api/v1/audits/{audit_id}": {
      "get": {
        "operationId": "get_audit_endpoint_api_v1_audits__audit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Endpoint"
      }
    },
    "/api/v1/audits/{audit_id}/answers": {
      "post": {
        "operationId": "answer_question_api_v1_audits__audit_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Question"
      }
    },
    "/api/v1/audits/{audit_id}/events": {
      "get": {
        "operationId": "audit_events_api_v1_audits__audit_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_AuditEvent__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Events"
      }
    },
    "/api/v1/audits/{audit_id}/repair-bundle": {
      "post": {
        "operationId": "create_repair_bundle_api_v1_audits__audit_id__repair_bundle_post",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_RepairBundleResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Repair Bundle"
      }
    },
    "/api/v1/audits/{audit_id}/repair-bundle/{bundle_id}/download": {
      "get": {
        "operationId": "download_repair_bundle_api_v1_audits__audit_id__repair_bundle__bundle_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "bundle_id",
            "required": true,
            "schema": {
              "title": "Bundle Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Repair Bundle"
      }
    },
    "/api/v1/audits/{audit_id}/report": {
      "post": {
        "operationId": "create_report_api_v1_audits__audit_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "audit_id",
            "required": true,
            "schema": {
              "title": "Audit Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_AuditReport_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Report"
      }
    },
    "/api/v1/benchmarks": {
      "get": {
        "operationId": "list_benchmarks_api_v1_benchmarks_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_BenchmarkDefinition__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Benchmarks"
      }
    },
...
  }
}